pass AR ID config file path as argument to main processing code
Prior to this change, the AR ID configuration hjson file was read within the main processing scripts (IVT calc, IVT at percentiles calc, AR ID). Now the path to the config file is passed as an argument to these scripts. This makes it easier to run multiple instances of the same code in parallel with different datasets on the same system -- different versions of the config file can be created and passed to the scripts at runtime, rather than assuming the location of a single config file.